Add Travel Service

+++ examples/train_ticket/workflow/travel/data.go
@@ -1,16 +1,16 @@
 package travel
 
 type TripResponse struct {
-	ComfortClass         uint16
-	EconomyClass         uint16
+	ComfortClass         int64
+	EconomyClass         int64
 	StartingStation      string
 	EndStation           string
 	StartingTime         string
 	EndTime              string
 	TripId               string
 	TrainTypeId          string
-	PriceForComfortClass float32
-	PriceForEconomyClass float32
+	PriceForComfortClass float64
+	PriceForEconomyClass float64
 
 	StopStations                  []string
 	NumberOfRestTicketFirstClass  uint16
